% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-clean-data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans source data from NOAA}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A data frame with raw data extracted from NOAA website}
}
\value{
A data frame with cleaned date, latitude, longitude and location
columns
}
\description{
Cleans source data from NOAA
}
\examples{
\dontrun{
df <- readr::read_delim("exdata/signif.txt", delim = "\\t")
library(dplyr)
library(stringr)
library(lubridate)
clean_data <- eq_clean_data(df)
}
}
