% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates a label for leaflet map}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data frame containing cleaned NOAA earthquake data}
}
\value{
A character vector with labels
}
\description{
Creates a label for leaflet map
}
\examples{
\dontrun{
df <- readr::read_delim("exdata/signif.txt", delim = "\\t")
eq_create_label(df)
}
}
