% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-map-state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots a map showing locations of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an interger representing a state.}

\item{year}{an interger representing a year}
}
\value{
A map that gives the lattitude and lognitude of accidents in a
selected  year.
}
\description{
Plot the exact location of the accident given the state number in a
particular year. This function stops if state.num is not a valid state id.
If there are no accidents, a message is printed as "no accidents to plot"
to the console.
}
\examples{
\dontrun{
fars_map_state(1, 2014)
}
}
