% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-read.R
\name{fars_read}
\alias{fars_read}
\title{Read in a data set}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a text with ".csv" or ".csv.bz2" as an extension}
}
\value{
a dataframe of raw data
}
\description{
Reads a data file from the current working directory and converts it to a
dataframe if a data file exist. Returns an error if the file doesn't exists, and
show an error message "file (file name) does not exist"
}
\examples{
\dontrun{
fars_read("data/accident_2013.csv.bz2")
}
}
