% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-read-years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a list of separate datasets by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector containing a number of years}
}
\value{
A list of data frames.
}
\description{
Reads a data file from the current working directory with a year and create
a dataframe with selected month and year. Returns a warning for every specified year for which
doesn't exist a matching data file showing what year the "invalid year" is.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}
}
