% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-summarize-years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Counts the total accidents in each month and each year.
Create a summury table of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an interger representing a year}
}
\description{
Counts the total accidents in each month and each year.
Create a summury table of accidents by month and year
}
\examples{
\donttest{
fars_summarize_years(c(2013, 2014, 2015))
}
}
