% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a character string representing a year}
}
\value{
A character string matching the complete filename
}
\description{
Make a new file name reflecting a given year
}
\examples{
\donttest{
make_filename(2014)
}
}
