% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for better timeline visualization in ggplot2}
\usage{
theme_timeline()
}
\description{
Theme for better timeline visualization in ggplot2
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
filter(country \%in\% c("USA", "MEXICO"), year > 2000) \%>\%
   ggplot(aes(x = datatime,
              y = country,
              color = as.numeric(deaths),
              size = as.numeric(eq_primary)
   )) +
   geom_timeline() +
   theme_timeline()
}

}
